+++ apps/examples/blinkid-preload/CHANGELOG.md
@@ -1,5 +1,14 @@
 # example-blinkid-preload
 
+## 1.1.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @microblink/camera-manager@7.2.1
+  - @microblink/blinkid-ux-manager@7.3.1
+  - @microblink/blinkid-core@7.3.1
+
 ## 1.1.7
 
 ### Patch Changes