+++ packages/blinkid-wasm/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @microblink/blinkid-wasm
 
+## 7.3.1
+
+### Minor Changes
+
+- Bumped version
+
 ## 7.3.0
 
 ### Minor Changes